fix update.sh

+++ update.sh
@@ -3,25 +3,34 @@ set -e
 
 cd "$(dirname "$(readlink -f "$BASH_SOURCE")")"
 
-versions=( "$@" )
-if [ ${#versions[@]} -eq 0 ]; then
-	versions=( */ )
+variants=( "$@" )
+if [ ${#variants[@]} -eq 0 ]; then
+	variants=$(ls -d -1 */*)
 fi
-versions=( "${versions[@]%/}" )
+variants=( "${variants[@]%/}" )
 
 
-for version in "${versions[@]}"; do
-	flavor="${version%%-*}" # "openjdk"
+for variant in ${variants[@]}; do
+	echo "### $variant ###"
+	[ -d "$variant" ] || continue;
+	echo -n "Checking $variant ... "
+
+	version="${variant%%/*}"
+	flavor="${version%%-*}" # "openjdk" or "java"
 	javaVersion="${version#*-}" # "6-jdk"
 	javaType="${javaVersion##*-}" # "jdk"
 	javaVersion="${javaVersion%-*}" # "6"
 	
-	dist="$(grep '^FROM ' "$version/Dockerfile" | cut -d' ' -f2)"
+	dist="$(grep '^FROM ' "$variant/Dockerfile" | cut -d' ' -f2)"
 	
 	fullVersion=
 	case "$flavor" in
 		openjdk)
 			debianVersion="$(set -x; docker run --rm "$dist" bash -c "apt-get update &> /dev/null && apt-cache show $flavor-$javaVersion-$javaType | grep '^Version: ' | head -1 | cut -d' ' -f2")"
+			fullVersion="${debianVersion%%-*}"
+			;;
+		java)
+			debianVersion="$(set -x; docker run --rm "$dist" bash -c "echo 'deb http://ppa.launchpad.net/webupd8team/java/ubuntu devel main' > /etc/apt/sources.list.d/webupd8team-ubuntu-java.list && apt-get update &> /dev/null && apt-cache show oracle-java$javaVersion-installer | grep '^Version: ' | head -1 | cut -d' ' -f2")"
 			fullVersion="${debianVersion%%-*}"
 			;;
 	esac
@@ -30,9 +39,9 @@ for version in "${versions[@]}"; do
 		(
 			set -x
 			sed -ri '
-				s/(ENV JAVA_VERSION) .*/\1 '"$fullVersion"'/g;
-				s/(ENV JAVA_DEBIAN_VERSION) .*/\1 '"$debianVersion"'/g;
-			' "$version/Dockerfile"
+				s/(JAVA_VERSION)=[^ ]+/\1='"$fullVersion"'/g;
+				s/(JAVA_DEBIAN_VERSION)=[^ ]+/\1='"$debianVersion"'/g;
+			' "$variant/Dockerfile"
 		)
 	fi
 done
